requirements.txt
torch>=2.2.0
torchvision>=0.17.0
albumentations>=1.4.0
opencv-python>=4.9.0
numpy>=1.26.0
PyYAML>=6.0.1
transformers>=4.40.0
tqdm>=4.66.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
kagglehub>=0.1.0
matplotlib>=3.8.0
